# energy-consumption-monitoring/docs/api-spec.json
{
    "openapi": "3.0.3",
    "info": {
        "title": "Energy Consumption Monitoring API",
        "version": "1.0.0",
        "description": "校园智慧后勤能耗监测系统后端 API 文档。"
    },
    "paths": {
...
        "/api/analysis/comparison/": {
            "get": {
                "operationId": "api_analysis_comparison_retrieve",
                "description": "返回当前周期与环比、同比的总量与变化率。",
                "tags": [
                    "api"
                ],
                "security": [
                    {
                        "jwtAuth": []
                    }
                ],
                "responses": {
                    "200": {
                        "description": "No response body"
                    }
                }
            }
        },
        "/api/analysis/dashboard/": {
            "get": {
                "operationId": "api_analysis_dashboard_retrieve",
                "description": "返回能耗总览、覆盖率和告警统计。",
                "tags": [
                    "api"
                ],
                "security": [
                    {
                        "jwtAuth": []
                    }
                ],
                "responses": {
                    "200": {
                        "description": "No response body"
                    }
                }
            }
        },
        "/api/analysis/distribution/": {
            "get": {
                "operationId": "api_analysis_distribution_retrieve",
                "description": "按区域或能源类型返回能耗占比分布。",
                "tags": [
                    "api"
                ],
                "security": [
                    {
                        "jwtAuth": []
                    }
                ],
                "responses": {
                    "200": {
                        "description": "No response body"
                    }
                }
            }
        },
        "/api/analysis/forecast/": {
            "get": {
                "operationId": "api_analysis_forecast_retrieve",
                "description": "基于历史日总量给出 7/30 天线性预测结果。",
                "tags": [
                    "api"
                ],
                "security": [
                    {
                        "jwtAuth": []
                    }
                ],
                "responses": {
                    "200": {
                        "description": "No response body"
                    }
                }
            }
        },
        "/api/analysis/ranking/": {
            "get": {
                "operationId": "api_analysis_ranking_retrieve",
                "description": "按楼宇/房间/部门返回能耗 TopN 排名。",
                "tags": [
                    "api"
                ],
                "security": [
                    {
                        "jwtAuth": []
                    }
                ],
                "responses": {
                    "200": {
                        "description": "No response body"
                    }
                }
            }
        },
        "/api/analysis/trend/": {
            "get": {
                "operationId": "api_analysis_trend_retrieve",
                "description": "按日/月/年粒度返回趋势序列。",
                "tags": [
                    "api"
                ],
                "security": [
                    {
                        "jwtAuth": []
                    }
                ],
                "responses": {
                    "200": {
                        "description": "No response body"
                    }
                }
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "jwtAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    }
}